+++ src/commands/index.ts
@@ -171,6 +171,7 @@ import { remoteEfb } from './support/remoteEfb';
 import { brakeCheck } from './support/brakeCheck';
 import { spad } from './general/spad';
 import { devLanguages } from './general/devLanguages';
+import { simbridgeDebug } from './support/simbridgeDebug'
 
 const commands: BaseCommandDefinition[] = [
     typeCommand,
@@ -344,6 +345,7 @@ const commands: BaseCommandDefinition[] = [
     brakeCheck,
     spad,
     devLanguages,
+    simbridgeDebug,
 ];
 
 const commandsObject: { [k: string]: BaseCommandDefinition } = {};